feat: added support for v4 migration and refactored the code.
