+++ .babelrc
@@ -5,7 +5,7 @@
       "@babel/preset-env",
       {
         "targets": {
-          "node": 12.13
+          "node": 14.15
         },
         "forceAllTransforms": true
       }
