--- schemes.js
@@ -1,71 +1,0 @@
-'use strict';
-
-var _require = require('gatsby/graphql'),
-    GraphQLEnumType = _require.GraphQLEnumType;
-
-exports.ImageResizingBehavior = new GraphQLEnumType({
-  name: 'ImageResizingBehavior',
-  values: {
-    NO_CHANGE: {
-      value: ''
-    },
-    PAD: {
-      value: 'pad',
-      description: 'Same as default resizing, but adds padding so that generated image has the specified dimensions.'
-    },
-    CROP: {
-      value: 'crop',
-      description: 'Crop a part of the original image to match the specified size.'
-    },
-    FILL: {
-      value: 'fill',
-      description: 'Crop the image to the specified dimensions, if the original image is smaller than these dimensions, then the image will be upscaled.'
-    },
-    THUMB: {
-      value: 'thumb',
-      description: 'When used in association with the f parameter below, creates a thumbnail from the image based on a focus area.'
-    },
-    SCALE: {
-      value: 'scale',
-      description: 'Scale the image regardless of the original aspect ratio.'
-    }
-  }
-});
-exports.ImageCropFocusType = new GraphQLEnumType({
-  name: 'ContentstackImageCropFocus',
-  values: {
-    TOP: {
-      value: "top"
-    },
-    TOP_LEFT: {
-      value: "top_left"
-    },
-    TOP_RIGHT: {
-      value: "top_right"
-    },
-    BOTTOM: {
-      value: "bottom"
-    },
-    BOTTOM_RIGHT: {
-      value: "bottom_left"
-    },
-    BOTTOM_LEFT: {
-      value: "bottom_right"
-    },
-    RIGHT: {
-      value: "right"
-    },
-    LEFT: {
-      value: "left"
-    },
-    FACE: {
-      value: "face"
-    },
-    FACES: {
-      value: "faces"
-    },
-    CENTER: {
-      value: "center"
-    }
-  }
-});
