+++ src/pre-bootstrap.js
@@ -1,0 +1,8 @@
+'use strict';
+
+exports.onPreBootstrap = ({ reporter }) => {
+  const args = process.argv;
+  if (args.includes('--verbose')) {
+    reporter.setVerbose(true);
+  }
+};
