+++ src/utils.js
@@ -41,6 +41,7 @@ exports.checkIfUnsupportedFormat = data => {
 
 exports.SUPPORTED_FILES_COUNT = 'SUPPORTED_FILES_COUNT';
 exports.IMAGE_REGEXP = new RegExp('https://(stag-images|(eu-)?images).(blz-)?contentstack.(io|com)/v3/assets/');
+exports.ASSET_NODE_UIDS = 'ASSET_NODE_UIDS';
 
 exports.CODES = {
   SyncError: '10001',
